Never assign the TRACKARTIST role
I've found over many years that I need to tag ALBUMARTIST to help LMS
(and other music players) recognize a single album. At the same time,
historically I've had a lot of trouble with the COMPILATION tag, so I
basically never set it. This means a LOT of my artists are getting
tagged TRACKARTIST, which is not what I want -- if I put someone in the
ARTIST tag, I want them to be browseable in the library.  The simplest
solution here appears to be remove the logic that converts ARTISTs to
TRACKARTISTs.