Add preference for enabling TrackArtist behavior
Per [1], when AlbumArtist is set but Compilation is false, contributors
in the Artist tag that are not in AlbumArtist are deemed "track artists"
and not included in the Artists menu when browsing.  Add a preference
that allows the user to control this behavior, giving them the choice to
add all artists to the browsing menu.

[1] https://wiki.slimdevices.com/index.php/Compilations.html#A_typical_setup